Attempt first solution for 1. Two Sum
This solution ended being the brute force solution for the problem. The
time complexity on this solution was O(n^2) and the space complexity
was O(1). Lookup times were expensive, so that could be improved using
a hash map to lookup the complement of the value instead of traversing
the array twice.