Attempt first solution for #26
Solution #1
When first tackling this problem, I had attempted to bubble numbers
from the front to the back. The first solution would take in the
current value of the iterator, and count up to the next instance of a
unique number. All numbers inclusive of the second and last appearance
of the current value would be converted to the next unique value. This
became a problem since I was creating exponential time complexity in
handling the same indexes multiple times. Although it worked, the
program timed out and was not accepted.

Solution #2
My next solution was simpler; I deleted values for which subsequent
occurences of the current values occured. Using `vector.erase`, the
vector automatically re-creates itself when values in the middle are
removed. The trick to this is, though, is that the memory addresses of
the vector change post-erase. Any pointers to the left of the erase
will remain the same; any pointers to the right of the erase will
change. This lead to problems since I was iterating through the vector
using `iter+1`, which gets the memory address physically next to the
iterator.

When a vector is constructed, sequential memory addresses are allotted
to the vector. When the vector does its automatic memory management,
new memory addresses may no longer be sequential. Using the `++`
operator to iterate through the vector will move the iterator to the
next memory address in the vector instead.