Attempt 1 for 6. ZigZag Conversion
This attempt failed to compile. It kept getting runtime errors. The
base idea of it was similar to the solution provided in the discussion
boards for this problem. One thing that should be changed is the data
structure used. In this attempt, I used a 2D array which would cost
O(n^2) space complexity which is undesirable. Simply using a string to
append the letters as needed would be more efficient space-wise.