Attempt 1 at 66 Plus One
This was my first attempt at solving an algorithm design problem using
Python. This is most optimal but not really the cleanest code
python-wise. O(n) matches best case, but there's a lot of code that can
be combined here. Rather than manipulating the same list, I could have
converted the existing list to an int, add 1, and then recreate a list
off the new int.